+++ MaxMind.MinFraud/Response/Insights.cs
@@ -22,6 +22,20 @@ namespace MaxMind.MinFraud.Response
         public CreditCard CreditCard { get; internal set; } = new CreditCard();
 
         /// <summary>
+        /// This object contains information about the device that MaxMind
+        /// believes is associated with the IP address passed in the request.
+        /// </summary>
+        [JsonProperty("device")]
+        public Device Device { get; internal set; } = new Device();
+
+        /// <summary>
+        /// This object contains information about the email address passed in
+        ///  the request.
+        /// </summary>
+        [JsonProperty("email")]
+        public Email Email { get; internal set; } = new Email();
+
+        /// <summary>
         /// An object containing minFraud data related to the shipping address
         ///  used in the transaction.
         /// </summary>
@@ -41,8 +55,7 @@ namespace MaxMind.MinFraud.Response
         /// <returns>A string that represents the current object.</returns>
         public override string ToString()
         {
-            return
-                $"{base.ToString()}, IPAddress: {{{IPAddress}}}, CreditCard: {{{CreditCard}}}, ShippingAddress: {{{ShippingAddress}}}, BillingAddress: {{{BillingAddress}}}";
+            return $"{base.ToString()}, IPAddress: {IPAddress}, CreditCard: {CreditCard}, Device: {Device}, Email: {Email}, ShippingAddress: {ShippingAddress}, BillingAddress: {BillingAddress}";
         }
     }
 }